Move rid and arch helpers
